% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_table.R
\name{league_table}
\alias{league_table}
\title{Create the official league table}
\usage{
league_table(results, mtr = FALSE, criteria = c("GD", "GS", "W"))
}
\arguments{
\item{results}{The results of the played matches.}

\item{mtr}{logical. If true, the ranks of teams with equal number of points are determined by their mutual results.}

\item{criteria}{The order of criteria on which the table is made when teams have equal number of points. Possible criteria are the goal difference ("GD"), the number of goals scored ("GS") and the number of matches won ("W")}
}
\value{
A dataframe with the final standing of the league.
}
\description{
This function creates the official league table, given the match results.
}
\examples{
teams <- paste0("Team ", LETTERS[1:10])
results <- subset(merge(teams, teams), x != y)
colnames(results) <- c("home_team", "away_team")
rownames(results) <- NULL
results$home_score <- rpois(nrow(results), lambda = 1)
results$away_score <- rpois(nrow(results), lambda = 1)
league_table(results)
}
