% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFSC.R
\docType{package}
\name{PFSC}
\alias{PFSC}
\title{Probabilistic final standing calculator}
\description{
Functions to create the probabilistic final standings of prematurely stopped leagues.
The main functions are \code{\link{league_table}}, \code{\link{team_ratings}} and \code{\link{simulate_league}}
}
\references{
Van Eetvelde, H., Hvattum, L. M., & Ley, C. (2021). The Probabilistic Final Standing Calculator: a fair stochastic tool to handle abruptly stopped football seasons. arXiv preprint arXiv:2101.10597.
}
\author{
Hans Van Eetvelde <Hans.VanEetvelde@UGent.be>

Christophe Ley

Lars Magnus Hvattum
}
