% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_league.R
\name{simulate_league}
\alias{simulate_league}
\title{Simulate the remainder of a double round robin league and create a probabilistic final standing.}
\usage{
simulate_league(
  results,
  S = 1e+05,
  mtr = FALSE,
  criteria = c("GD", "GS", "W"),
  strengths = FALSE
)
}
\arguments{
\item{results}{The results of the played matches.}

\item{mtr}{logical. If true, the ranks of teams with equal number of points is determined by their mutual results.}

\item{criteria}{The order of criteria on which the table is made when teams have equal number of points. Possible criteria are the goal difference ("GD"), the number of goals scored ("GS") and the number of matches won ("W")}

\item{strengths}{logical. If true, the function also returns the estimated team strengths}
}
\value{
A dataframe with the probabilistic final standing of the league.
}
\description{
This function creates a probabilistic final standing, given the played matches
}
\examples{
teams<-paste0("Team ",LETTERS[1:10])
results<-subset(merge(teams, teams), x!=y)
colnames(results)<-c("home_team","away_team")
rownames(results)<-NULL
results$home_score<-rpois(nrow(results),lambda = 1)
results$away_score<-rpois(nrow(results),lambda = 1)
simulate_league(results)
}
