% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFSC.R
\docType{package}
\name{PFSC}
\alias{PFSC}
\title{Probabilistic final standing calculator}
\description{
Functions to create the probabilistic final standings of prematurely stopped leagues.
}
\author{
Hans Van Eetvelde <Hans.VanEetvelde@UGent.be>
}
